Move Maker to separate module
