+++ src/lib.rs
@@ -24,6 +24,7 @@ pub mod ethereum_wallet;
 pub mod float_maths;
 pub mod geth;
 pub mod jsonrpc;
+pub mod maker;
 pub mod mid_market_rate;
 pub mod network;
 pub mod ongoing_swaps;
@@ -33,6 +34,7 @@ pub mod seed;
 pub mod swap;
 pub mod swap_id;
 
+pub use maker::Maker;
 pub use mid_market_rate::MidMarketRate;
 pub use rate::{Rate, Spread};
 pub use seed::Seed;
